requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cfgv==3.3.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.6
Django==4.1.7
django-appconf==1.0.5
django-filter==22.1
django-fsm==2.8.1
django-fsm-log==3.0.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
exceptiongroup==1.1.1
filelock==3.10.0
identify==2.5.21
idna==3.4
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.2
nodeenv==1.7.0
numpy==1.24.2
openpyxl==3.1.2
packaging==23.0
pandas==1.5.3
platformdirs==3.1.1
pluggy==1.0.0
pre-commit==3.2.0
prompt-toolkit==3.0.38
psycopg2==2.9.5
pytest==7.2.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
PyYAML==6.0
redis==4.5.1
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.15
vine==5.0.0
virtualenv==20.21.0
wcwidth==0.2.6
wheel==0.40.0
xlrd==2.0.1
xmltodict==0.13.0
